requirements.txt
pytrends
google_generativeai
# psycopg2
pandas
requests
